feat(client): vector catalogue
