+++ packages/client/src/components/engine/index.ts
@@ -1,4 +1,4 @@
-export * from "./EngineAccessButton";
 export * from "./edit-engine-details";
+export * from "./engine-access-button";
 export * from "./engine-header";
 export * from "./generic-engine-cards";
