+++ packages/client/src/components/settings/index.ts
@@ -1,9 +1,9 @@
 export * from "./EngineModelTestSidebar";
 export * from "./EngineQASidebar";
 export * from "./FileTable";
-export * from "./MembersTable";
-export * from "./PendingMembersTable";
-export * from "./SettingsTiles";
+export * from "./members-table";
+export * from "./pending-members-table";
 export * from "./settings.types";
+export * from "./settings-tiles";
 export * from "./UserTable";
 export * from "./update-smss";
