+++ packages/client/src/components/settings/members-add-overlay.tsx
@@ -1,0 +1,994 @@
+import type { AxiosResponse } from "axios";
+import { Edit, Eye, X } from "lucide-react";
+import { useEffect, useState } from "react";
+import { useDebouncedValue } from "@semoss/sdk/react";
+import {
+	Avatar,
+	AvatarFallback,
+	Button,
+	Card,
+	CardDescription,
+	CardHeader,
+	CardTitle,
+	Command,
+	CommandEmpty,
+	CommandGroup,
+	CommandInput,
+	CommandItem,
+	CommandList,
+	Dialog,
+	DialogContent,
+	DialogFooter,
+	DialogHeader,
+	DialogTitle,
+	Field,
+	Input,
+	Label,
+	P,
+	Popover,
+	PopoverContent,
+	PopoverTrigger,
+	RadioGroup,
+	RadioGroupItem,
+	Select,
+	SelectContent,
+	SelectItem,
+	SelectTrigger,
+	SelectValue,
+	Spinner,
+	toast,
+} from "@semoss/ui/next";
+import {
+	addEngineUserPermissions,
+	addProjectUserPermissions,
+	editEngineUserPermissions,
+	editProjectUserPermissions,
+	getEngineUsers,
+	getEngineUsersNoCredentials,
+	getProjectUsers,
+	getProjectUsersNoCredentials,
+} from "@/api";
+import { PERMISSION_DESCRIPTION_MAP } from "@/constants";
+import { useSettings } from "@/hooks";
+import type { ALL_TYPES } from "@/types";
+import { permissionPriorityMapper } from "@/utility/general";
+import { MembersAddOverlayUser } from "./members-add-overlayUser";
+import type { SETTINGS_ROLE } from "./settings.types";
+
+const Setting_Role_Values: SETTINGS_ROLE[] = ["Author", "Editor", "Read-Only"];
+
+const validSetting = (value: unknown) => {
+	return Setting_Role_Values.includes(value as SETTINGS_ROLE);
+};
+
+const AUTOCOMPLETE_OFFSET = 0;
+const AUTOCOMPLETE_LIMIT = 10;
+
+interface MembersAddOverlayProps {
+	/**
+	 * Type of engine
+	 */
+	type: ALL_TYPES;
+
+	/**
+	 * ID of the app or engine being edited
+	 */
+	id: string;
+
+	/**
+	 * Track if the model is open or close
+	 */
+	open: boolean;
+
+	/**
+	 * User we want to edit
+	 */
+	user?: User;
+
+	/**
+	 * Set Edit user to null
+	 *
+	 */
+	setAddModalUser?: React.Dispatch<React.SetStateAction<User>>;
+
+	/**
+	 * User permission of the app or engine being edited
+	 */
+	userPermission: SETTINGS_ROLE;
+
+	/**
+	 * Called on close
+	 *
+	 * @returns - method that is called onClose
+	 */
+	onClose: (success: boolean) => void;
+
+	/**
+	 * Called on close
+	 *
+	 * @returns - method that is called onClose
+	 */
+	onChange?: () => void;
+}
+
+interface User {
+	id: string;
+	type: string;
+	name: string;
+	email: string;
+	permission_granted_by_type: string;
+	permission_granted_by: string;
+	permission: string;
+	date_added: string;
+	usage_restriction?: string;
+	usage_frequency?: string;
+	max_tokens?: number;
+	max_response_time?: number;
+}
+
+export const MembersAddOverlay = (props: MembersAddOverlayProps) => {
+	const {
+		type,
+		id,
+		open = false,
+		userPermission,
+		onClose = () => null,
+		user,
+		setAddModalUser,
+		onChange = () => null,
+	} = props;
+	const { adminMode } = useSettings();
+
+	/** Add Member State */
+	const [selectedMembers, setSelectedMembers] = useState([]);
+	const [commandOpen, setCommandOpen] = useState(false);
+
+	const [selectedRole, setSelectedRole] = useState<SETTINGS_ROLE>(null);
+	const [search, setSearch] = useState<string>("");
+	const [restriction, setRestriction] = useState<string>("null");
+	const [maxTokens, setMaxTokens] = useState<string>("");
+	const [maxTime, setMaxTime] = useState<string>("");
+	const [frequency, setFrequency] = useState<string>("");
+
+	//modal member logic
+	const [isScrollBottom, setIsScrollBottom] = useState(false);
+	const [offset, setOffset] = useState(AUTOCOMPLETE_OFFSET);
+	const [renderedMembers, setRenderedMembers] = useState([]);
+	const [infiniteOn, setInfiniteOn] = useState(true);
+	const [searchLoading, setSearchLoading] = useState(false);
+
+	// debounce the input
+	const debouncedSearch = useDebouncedValue(search);
+
+	const usageRestritctionTypes: Record<string, string> = {
+		null: "None",
+		token: "Token",
+		compute: "Compute time",
+	};
+	const frequencyTypes: Record<string, string> = {
+		DAY: "Daily",
+		WEEK: "Weekly",
+		MONTH: "Monthly",
+	};
+	const unitTypes: string[] = ["milliseconds"];
+
+	useEffect(() => {
+		setSelectedRole("Read-Only");
+		if (user) {
+			setSelectedRole(
+				permissionPriorityMapper(user?.permission)
+					?.permission as SETTINGS_ROLE,
+			);
+			setRestriction(
+				user?.usage_restriction !== undefined
+					? user?.usage_restriction
+					: "null",
+			);
+			setMaxTokens(user?.max_tokens?.toString());
+			setMaxTime(user?.max_response_time?.toString());
+			setFrequency(user?.usage_frequency);
+		}
+
+		// reset on open or close
+		setSelectedMembers([]);
+		setSearch("");
+	}, [open]);
+
+	useEffect(() => {
+		if (!open) return;
+
+		const cancelled = false;
+		setSearchLoading(true);
+
+		const fetchUsers = async () => {
+			try {
+				let all = [];
+				if (type === "PROJECT") {
+					const [noCred, cred] = await Promise.all([
+						getProjectUsersNoCredentials(
+							adminMode,
+							id,
+							AUTOCOMPLETE_LIMIT,
+							offset,
+							debouncedSearch || "",
+						),
+						getProjectUsers(
+							adminMode,
+							id,
+							debouncedSearch || "",
+							"", // permission
+							offset,
+							AUTOCOMPLETE_LIMIT,
+						),
+					]);
+					all = [...(noCred?.data || []), ...(cred?.members || [])];
+				} else if (
+					type === "DATABASE" ||
+					type === "STORAGE" ||
+					type === "MODEL" ||
+					type === "VECTOR" ||
+					type === "FUNCTION"
+				) {
+					const [noCred, cred] = await Promise.all([
+						getEngineUsersNoCredentials(
+							adminMode,
+							id,
+							AUTOCOMPLETE_LIMIT,
+							offset,
+							debouncedSearch || "",
+						),
+						getEngineUsers(
+							adminMode,
+							id,
+							debouncedSearch || "",
+							"", // permission
+							offset,
+							AUTOCOMPLETE_LIMIT,
+						),
+					]);
+					all = [...(noCred?.data || []), ...(cred?.members || [])];
+				} else {
+					setSearchLoading(false);
+					return;
+				}
+
+				if (!cancelled) {
+					if (all.length < AUTOCOMPLETE_LIMIT) setInfiniteOn(false);
+					if (
+						renderedMembers.length >= AUTOCOMPLETE_LIMIT &&
+						offset > 0
+					) {
+						setRenderedMembers((prev) => [...prev, ...all]);
+					} else {
+						setRenderedMembers(all);
+					}
+					setSearchLoading(false);
+				}
+			} catch (e) {
+				if (!cancelled) {
+					toast.error(String(e));
+					setSearchLoading(false);
+				}
+			}
+		};
+
+		fetchUsers();
+	}, [open, debouncedSearch, offset, adminMode, id, type]);
+
+	const isLoading = searchLoading;
+
+	const getAdditionalMembers = () => {
+		setOffset(offset + AUTOCOMPLETE_LIMIT);
+	};
+
+	/**
+	 * Update the selected users
+	 * @param members
+	 * @param quickUpdate
+	 * @returns
+	 */
+	const updateUser = async (members) => {
+		let success = false;
+		try {
+			// construct requests for post data
+			const requests = members.map((m) => {
+				let json: Record<string, unknown> = {
+					userid: m.id,
+					permission: validSetting(selectedRole)
+						? permissionPriorityMapper(selectedRole)?.permission
+						: selectedRole,
+				};
+
+				// FOR MODELS
+				if (restriction !== "null") {
+					json = {
+						...json,
+						usageRestriction: restriction,
+					};
+				}
+
+				if (frequency) {
+					json = {
+						...json,
+						usageFrequency: frequency,
+					};
+				}
+
+				if (restriction === "token") {
+					json = {
+						...json,
+						maxTokens: Number(maxTokens),
+					};
+				}
+
+				if (restriction === "compute") {
+					json = {
+						...json,
+						maxResponseTime: Number(maxTime),
+					};
+				}
+				return json;
+			});
+
+			if (requests.length === 0) {
+				toast.warning("No permissions to change");
+				return;
+			}
+
+			let response:
+				| AxiosResponse<{ success: boolean }>
+				| {
+						response: Response;
+						data: {
+							success: boolean;
+						};
+				  }
+				| null = null;
+			if (
+				type === "DATABASE" ||
+				type === "STORAGE" ||
+				type === "MODEL" ||
+				type === "VECTOR" ||
+				type === "FUNCTION"
+			) {
+				response = await editEngineUserPermissions(
+					adminMode,
+					id,
+					requests,
+				);
+			} else if (type === "PROJECT") {
+				response = await editProjectUserPermissions(
+					adminMode,
+					id,
+					requests,
+				);
+			}
+
+			if (!response) {
+				return;
+			}
+
+			// ignore if there is no response
+			if (response.data.success) {
+				toast.success("Successfully updated user permissions");
+				success = true;
+				onChange();
+			} else {
+				toast.error("Error changing user permissions");
+			}
+		} catch (e) {
+			toast.error(String(e));
+		} finally {
+			closeOverlay(type, success);
+		}
+	};
+
+	/**
+	 * @name addMembers
+	 *
+	 * Add members to the app or engine
+	 */
+	const addMembers = async () => {
+		let success = false;
+
+		try {
+			// construct requests for post data
+			let requests: unknown[] = [];
+			if (type === "MODEL") {
+				requests = selectedMembers.map((m) => {
+					return {
+						userid: m.id,
+						permission:
+							permissionPriorityMapper(selectedRole)?.permission,
+						email: m.email,
+						name: m.name,
+						type: m.type,
+						username: m.username,
+						usageRestriction:
+							restriction === "null" ? null : restriction,
+						usageFrequency: frequency,
+						...(restriction === "token" && {
+							maxTokens: Number(maxTokens),
+						}),
+						...(restriction === "compute" && {
+							maxResponseTime: Number(maxTime),
+						}),
+					};
+				});
+			} else {
+				requests = selectedMembers.map((m) => {
+					return {
+						userid: m.id,
+						permission:
+							permissionPriorityMapper(selectedRole)?.permission,
+						email: m.email,
+						name: m.name,
+						type: m.type,
+						username: m.username,
+					};
+				});
+			}
+
+			if (requests.length === 0) {
+				toast.warning("No permissions to change");
+				return;
+			}
+
+			let response:
+				| AxiosResponse<{ success: boolean }>
+				| {
+						response: Response;
+						data: {
+							success: boolean;
+						};
+				  }
+				| null = null;
+			if (
+				type === "DATABASE" ||
+				type === "STORAGE" ||
+				type === "MODEL" ||
+				type === "VECTOR" ||
+				type === "FUNCTION"
+			) {
+				response = await addEngineUserPermissions(
+					adminMode,
+					id,
+					requests,
+				);
+			} else if (type === "PROJECT") {
+				response = await addProjectUserPermissions(
+					adminMode,
+					id,
+					requests as string[],
+				);
+			}
+
+			if (!response) {
+				return;
+			}
+
+			// ignore if there is no response
+			if (response.data.success) {
+				toast.success("Successfully added member permissions");
+				success = true;
+			} else {
+				toast.error("Error changing user permissions");
+			}
+		} catch (e) {
+			toast.error(String(e));
+		} finally {
+			// close the overlay
+			closeOverlay(type, success);
+		}
+	};
+
+	const nearBottom = (
+		target: {
+			scrollHeight?: number;
+			scrollTop?: number;
+			clientHeight?: number;
+		} = {},
+	) => {
+		const diff = Math.round(target.scrollHeight - target.scrollTop);
+		return diff - 25 <= target.clientHeight;
+	};
+
+	const closeOverlay = (type: ALL_TYPES, isSuccess: boolean) => {
+		if (type === "MODEL") {
+			setRestriction("null");
+			setFrequency("");
+			setMaxTime("");
+			setMaxTokens("");
+		}
+		setAddModalUser(null);
+		onClose(isSuccess);
+	};
+
+	useEffect(() => {
+		if (isScrollBottom) {
+			if (infiniteOn) {
+				getAdditionalMembers();
+			}
+		}
+	}, [isScrollBottom]);
+
+	const removeMember = (userId: string) => {
+		const filtered = selectedMembers.filter((val) => val.id !== userId);
+		setSelectedMembers(filtered);
+	};
+
+	return (
+		<Dialog open={open} onOpenChange={() => closeOverlay(type, false)}>
+			<DialogContent
+				className="flex h-fit max-h-[90vh] flex-col gap-4 overflow-auto sm:max-w-2xl"
+				data-testid="members-add-overlay-modal"
+			>
+				<DialogHeader>
+					<DialogTitle data-testid="members-add-overlay-modal-title">
+						{user === null ? "Add Members" : "Edit Member"}
+					</DialogTitle>
+				</DialogHeader>
+
+				<div className="flex flex-col gap-4">
+					{user === null && (
+						<Popover
+							open={commandOpen}
+							onOpenChange={setCommandOpen}
+						>
+							<PopoverTrigger asChild>
+								<Button
+									variant="outline"
+									role="combobox"
+									className="w-full justify-between"
+									data-testid="members-add-overlay-autocomplete"
+								>
+									{selectedMembers.length === 0
+										? "Search users"
+										: selectedMembers.length === 1
+											? selectedMembers[0].name
+											: `${selectedMembers.length} members selected`}
+								</Button>
+							</PopoverTrigger>
+							<PopoverContent className="w-[600px] p-0">
+								<Command shouldFilter={false}>
+									<CommandInput
+										placeholder="Search users..."
+										value={search}
+										onValueChange={(value) => {
+											setSearch(value);
+											setOffset(0);
+											setInfiniteOn(true);
+											setRenderedMembers([]);
+											setSearchLoading(true);
+										}}
+									/>
+									<CommandList
+										onScroll={({ currentTarget }) => {
+											setIsScrollBottom(
+												nearBottom({
+													scrollHeight:
+														currentTarget.scrollHeight,
+													scrollTop:
+														currentTarget.scrollTop,
+													clientHeight:
+														currentTarget.clientHeight,
+												}),
+											);
+										}}
+									>
+										{isLoading ? (
+											<div className="flex items-center justify-center p-4">
+												<Spinner />
+											</div>
+										) : (
+											<>
+												<CommandEmpty>
+													No users found.
+												</CommandEmpty>
+												<CommandGroup>
+													{renderedMembers.map(
+														(option) => {
+															const hasPermission =
+																!!option.permission;
+															const isSelected =
+																selectedMembers.some(
+																	(m) =>
+																		m.id ===
+																		option.id,
+																);
+															return (
+																<CommandItem
+																	key={
+																		option.id
+																	}
+																	disabled={
+																		hasPermission
+																	}
+																	onSelect={() => {
+																		if (
+																			!hasPermission
+																		) {
+																			if (
+																				isSelected
+																			) {
+																				removeMember(
+																					option.id,
+																				);
+																			} else {
+																				setSelectedMembers(
+																					[
+																						...selectedMembers,
+																						option,
+																					],
+																				);
+																			}
+																		}
+																	}}
+																	className="justify-between"
+																>
+																	<div className="max-w-[85%] overflow-hidden">
+																		<MembersAddOverlayUser
+																			name={
+																				option.name
+																			}
+																			id={
+																				option.id
+																			}
+																			email={
+																				option.email
+																			}
+																			type={
+																				option.type
+																			}
+																		/>
+																	</div>
+																	{hasPermission && (
+																		<span className="whitespace-nowrap text-muted-foreground text-xs">
+																			Already
+																			Added
+																		</span>
+																	)}
+																	{isSelected &&
+																		!hasPermission && (
+																			<span className="text-primary text-xs">
+																				✓
+																			</span>
+																		)}
+																</CommandItem>
+															);
+														},
+													)}
+												</CommandGroup>
+											</>
+										)}
+									</CommandList>
+								</Command>
+							</PopoverContent>
+						</Popover>
+					)}
+
+					<div
+						className="flex max-h-[200px] flex-col gap-4 overflow-auto"
+						data-testid="members-add-overlay-outerbox"
+					>
+						{user === null &&
+							selectedMembers.map((selectedUser) => (
+								<MembersAddOverlayUser
+									key={selectedUser.id}
+									name={selectedUser.name}
+									id={selectedUser.id}
+									email={selectedUser.email}
+									type={selectedUser.type}
+									action={
+										<Button
+											variant="ghost"
+											size="icon"
+											onClick={() => {
+												removeMember(selectedUser.id);
+											}}
+										>
+											<X className="h-4 w-4" />
+										</Button>
+									}
+								/>
+							))}
+
+						{user !== null && (
+							<MembersAddOverlayUser
+								key={user.id}
+								name={user.name}
+								id={user.id}
+								email={user.email}
+								type={user.type}
+							/>
+						)}
+					</div>
+
+					<div className="flex flex-col gap-2">
+						<P
+							className="font-medium"
+							data-testid="members-permissions"
+						>
+							Permissions
+						</P>
+						<div className="rounded-lg bg-muted/30 p-4">
+							<RadioGroup
+								value={selectedRole}
+								onValueChange={(val) => {
+									if (val) {
+										setSelectedRole(val as SETTINGS_ROLE);
+									}
+								}}
+								className="flex flex-col gap-4"
+							>
+								<Card className="rounded-xl">
+									<CardHeader className="pb-3">
+										<div className="flex items-center justify-between">
+											<div className="flex items-center gap-3">
+												<Avatar className="h-5 w-5 font-bold text-xs">
+													<AvatarFallback className="bg-black/50 text-white">
+														A
+													</AvatarFallback>
+												</Avatar>
+												<div className="flex flex-col">
+													<CardTitle
+														className="text-base"
+														data-testid="author-role"
+													>
+														Author
+													</CardTitle>
+													<CardDescription className="text-sm">
+														{PERMISSION_DESCRIPTION_MAP[
+															type
+														].author ||
+															`Error: update key in test-editor.constants to "${name}"`}
+													</CardDescription>
+												</div>
+											</div>
+											<RadioGroupItem
+												value="Author"
+												disabled={
+													!adminMode &&
+													permissionPriorityMapper(
+														userPermission,
+													)?.priority > 1
+												}
+												data-testid="author-role-radio"
+											/>
+										</div>
+									</CardHeader>
+								</Card>
+
+								<Card className="rounded-xl">
+									<CardHeader className="pb-3">
+										<div className="flex items-center justify-between">
+											<div className="flex items-center gap-3">
+												<div className="flex h-6 w-6 items-center justify-center text-black/50">
+													<Edit className="h-5 w-5" />
+												</div>
+												<div className="flex flex-col">
+													<CardTitle
+														className="text-base"
+														data-testid="editor-role"
+													>
+														Editor
+													</CardTitle>
+													<CardDescription className="text-sm">
+														{PERMISSION_DESCRIPTION_MAP[
+															type
+														].editor ||
+															`Error: update key in test-editor.constants to "${name}"`}
+													</CardDescription>
+												</div>
+											</div>
+											<RadioGroupItem
+												value="Editor"
+												disabled={
+													!adminMode &&
+													permissionPriorityMapper(
+														userPermission,
+													)?.priority > 2
+												}
+												data-testid="editor-role-radio"
+											/>
+										</div>
+									</CardHeader>
+								</Card>
+
+								<Card className="rounded-xl">
+									<CardHeader className="pb-3">
+										<div className="flex items-center justify-between">
+											<div className="flex items-center gap-3">
+												<div className="flex h-6 w-6 items-center justify-center text-black/50">
+													<Eye className="h-5 w-5" />
+												</div>
+												<div className="flex flex-col">
+													<CardTitle
+														className="text-base"
+														data-testid="readonly-role"
+													>
+														Read-Only
+													</CardTitle>
+													<CardDescription className="text-sm">
+														{PERMISSION_DESCRIPTION_MAP[
+															type
+														].readonly ||
+															`Error: update key in test-editor.constants to "${name}"`}
+													</CardDescription>
+												</div>
+											</div>
+											<RadioGroupItem
+												value="Read-Only"
+												disabled={
+													!adminMode &&
+													permissionPriorityMapper(
+														userPermission,
+													)?.priority > 3
+												}
+												data-testid="readonly-role-radio"
+											/>
+										</div>
+									</CardHeader>
+								</Card>
+							</RadioGroup>
+						</div>
+					</div>
+
+					{type === "MODEL" && (
+						<>
+							<P
+								className="font-medium"
+								data-testid="model-limit-restrictions"
+							>
+								Model Limit Restrictions
+							</P>
+							<div className="flex flex-col gap-4">
+								<Field>
+									<Label>Limit Type</Label>
+									<Select
+										value={restriction}
+										onValueChange={(value) => {
+											setRestriction(value);
+										}}
+									>
+										<SelectTrigger data-testid="model-limit-restrictions-select">
+											<SelectValue />
+										</SelectTrigger>
+										<SelectContent>
+											{Object.entries(
+												usageRestritctionTypes,
+											).map((option) => {
+												return (
+													<SelectItem
+														value={option[0]}
+														key={`usageRestrictionType-${option[0]}`}
+													>
+														{option[1]}
+													</SelectItem>
+												);
+											})}
+										</SelectContent>
+									</Select>
+								</Field>
+								{restriction === "token" && (
+									<Field>
+										<Label>Max Tokens</Label>
+										<Input
+											value={maxTokens}
+											type="number"
+											onChange={(e) => {
+												setMaxTokens(e.target.value);
+											}}
+											data-testid="model-max-tokens"
+										/>
+									</Field>
+								)}
+								{restriction === "compute" && (
+									<div className="flex gap-4">
+										<Field className="flex-1">
+											<Label>Max Response Time</Label>
+											<Input
+												value={maxTime}
+												type="number"
+												onChange={(e) => {
+													setMaxTime(e.target.value);
+												}}
+												data-testid="model-max-response-time"
+											/>
+										</Field>
+										<Field className="w-40">
+											<Label>Unit</Label>
+											<Select value={unitTypes[0]}>
+												<SelectTrigger>
+													<SelectValue />
+												</SelectTrigger>
+												<SelectContent>
+													{unitTypes.map((option) => {
+														return (
+															<SelectItem
+																value={option}
+																key={`unitType-${option}`}
+															>
+																{option}
+															</SelectItem>
+														);
+													})}
+												</SelectContent>
+											</Select>
+										</Field>
+									</div>
+								)}
+								{restriction !== "null" && (
+									<Field>
+										<Label>Frequency</Label>
+										<Select
+											value={frequency}
+											onValueChange={(value) => {
+												setFrequency(value);
+											}}
+										>
+											<SelectTrigger data-testid="model-frequency-select">
+												<SelectValue />
+											</SelectTrigger>
+											<SelectContent>
+												{Object.entries(
+													frequencyTypes,
+												).map((option) => {
+													return (
+														<SelectItem
+															value={option[0]}
+															key={`frequencyType-${option[0]}`}
+														>
+															{option[1]}
+														</SelectItem>
+													);
+												})}
+											</SelectContent>
+										</Select>
+									</Field>
+								)}
+							</div>
+						</>
+					)}
+				</div>
+
+				<DialogFooter>
+					<Button
+						variant="outline"
+						onClick={() => closeOverlay(type, false)}
+						data-testid="members-add-overlay-cancel-button"
+					>
+						Cancel
+					</Button>
+					{user === null && (
+						<Button
+							variant="default"
+							disabled={
+								!selectedRole || selectedMembers.length < 1
+							}
+							onClick={() => {
+								addMembers();
+							}}
+							data-testid="members-add-overlay-add-button"
+						>
+							Save
+						</Button>
+					)}
+
+					{user !== null && (
+						<Button
+							variant="default"
+							disabled={!selectedRole}
+							onClick={() => {
+								updateUser([user]);
+							}}
+							data-testid="members-add-overlay-update-button"
+						>
+							Update
+						</Button>
+					)}
+				</DialogFooter>
+			</DialogContent>
+		</Dialog>
+	);
+};
