+++ packages/client/src/pages/engine/engine-settingsPage.tsx
@@ -1,22 +1,12 @@
 import { useNavigate } from "react-router-dom";
-import { styled } from "@semoss/ui";
 import {
 	MembersTable,
 	PendingMembersTable,
 	SettingsTiles,
 } from "@/components/settings";
-import { TeamsTable } from "@/components/settings/TeamsTable";
+import { TeamsTable } from "@/components/settings/teams-table";
 import { SettingsContext } from "@/contexts";
 import { useEngine } from "@/hooks";
-
-const StyledContainer = styled("div")(({ theme }) => ({
-	width: "100%",
-	display: "flex",
-	alignSelf: "stretch",
-	flexDirection: "column",
-	alignItems: "flex-start",
-	gap: theme.spacing(3),
-}));
 
 export const EngineSettingsPage = () => {
 	const { name, path, type, active } = useEngine();
@@ -28,7 +18,7 @@ export const EngineSettingsPage = () => {
 				adminMode: false,
 			}}
 		>
-			<StyledContainer>
+			<div className="flex w-full flex-col items-start gap-6 self-stretch">
 				<SettingsTiles
 					type={type}
 					id={active.id}
@@ -40,7 +30,7 @@ export const EngineSettingsPage = () => {
 				/>
 				<PendingMembersTable type={type} id={active.id} />
 				<MembersTable type={type} id={active.id} />
-			</StyledContainer>
+			</div>
 			<div style={{ marginTop: 24 }}>
 				<TeamsTable type="ENGINE" id={active.id} />
 			</div>
