+++ packages/client/src/pages/engine/engine.constants.ts
@@ -8,13 +8,13 @@ import { ModelBrain } from "@/assets/img/ModelBrain";
 import type { ENGINE_TYPES, Role } from "@/types";
 import { EngineFilePage } from "./EngineFilePage";
 import { EngineQAPage } from "./EngineQAPage";
-import { EngineSettingsPage } from "./EngineSettingsPage";
 import { EngineSmssPage } from "./EngineSmssPage";
 import { EngineFileManagerPage } from "./engine-file-manager-page";
 import { EngineMetadataPage } from "./engine-metadata-page";
 import { EngineModelChatPage } from "./engine-model-chat-page";
 import { EngineOverviewPage } from "./engine-overview-page";
 import { EngineQueryDataPage } from "./engine-query-data-page";
+import { EngineSettingsPage } from "./engine-settingsPage";
 import { EngineUsagePage } from "./engine-usage-page";
 
 export const ENGINE_ROUTES: {
