+++ packages/client/src/pages/import/index.ts
@@ -1,6 +1,6 @@
 export * from "./EstablishConnectionPage";
 export * from "./ImportConnectionPage";
 export * from "./ImportLayout";
-export * from "./ImportPage";
 export * from "./ImportPageContent";
 export * from "./import.constants";
+export * from "./import-page";
